nlp-speech/requirements.txt
openai-whisper
torch
gTTS
soundfile
